pyproject.toml
[project]
name = "fastapi-suren"
version = "0.1.0"
description = ""
authors = [
    {name = "VincentVisage",email = "144054905+VincentVisage@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standart] (>=0.34.2,<0.35.0)",
    "pydantic[email] (>=2.11.4,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.40,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
